Switched core data type to scalar, API to float. scalar can be set in the root CMakeLists.txt to be e.g. float or double. The goal is to be able to use float, e.g. when running on a graphics card in order to lower memory usage and to speed up calculations. The API uses float for everything in order to provide a sensible interface. For the functions which get data pointers (e.g. System_Get_Spin_Directions), scalar should be set appropriately for static wrapper layers (see python core module). Renamed Version.h to Core_Defines.h and added typedef of scalar. Modified UIs accordingly. NOTE: ui-web, ui-python, ui-julia are untested and the CMakeLists should probably be adapted.

+++ core/include/interface/Interface_Chain.h
@@ -24,10 +24,10 @@ DLLEXPORT void Chain_Replace_Image(State * state, int idx_image=-1, int idx_chai
 DLLEXPORT bool Chain_Delete_Image(State * state, int idx_image=-1, int idx_chain=-1);
 
 // Get Data
-std::vector<double> Chain_Get_Rx(State * state, int idx_chain=-1);
-std::vector<double> Chain_Get_Rx_Interpolated(State * state, int idx_chain = -1);
-std::vector<double> Chain_Get_Energy_Interpolated(State * state, int idx_chain = -1);
-std::vector<std::vector<double>> Chain_Get_Energy_Array_Interpolated(State * state, int idx_chain=-1);
+std::vector<float> Chain_Get_Rx(State * state, int idx_chain=-1);
+std::vector<float> Chain_Get_Rx_Interpolated(State * state, int idx_chain = -1);
+std::vector<float> Chain_Get_Energy_Interpolated(State * state, int idx_chain = -1);
+std::vector<std::vector<float>> Chain_Get_Energy_Array_Interpolated(State * state, int idx_chain=-1);
 
 // Update Data (primarily for plots)
 DLLEXPORT void Chain_Update_Data(State * state, int idx_chain=-1);
