+++ ui-python/core/transition.py
@@ -14,7 +14,7 @@ def Homogeneous(p_state, idx_1, idx_2, idx_chain=-1):
 
 ### Add some temperature-scaled noise to a transition between two images of a chain
 _Add_Noise_Temperature             = _core.Transition_Add_Noise_Temperature
-_Add_Noise_Temperature.argtypes    = [ctypes.c_void_p, ctypes.c_double, ctypes.c_int, ctypes.c_int, ctypes.c_int]
+_Add_Noise_Temperature.argtypes    = [ctypes.c_void_p, ctypes.c_float, ctypes.c_int, ctypes.c_int, ctypes.c_int]
 _Add_Noise_Temperature.restype     = None
 def Add_Noise_Temperature(p_state, temperature, idx_1, idx_2, idx_chain=-1):
     _Add_Noise_Temperature(p_state, temperature, idx_1, idx_2, idx_chain)
